Falta controller de Estadisicas

+++ src/main/java/com/jerezsurinmobiliaria/web/dto/InmuebleDashboardDTO.java
@@ -9,6 +9,6 @@ public class InmuebleDashboardDTO {
     private String tipoVivienda;
     private String direccion;
     private Double precio;
-    private Long numeroInteresados;
+    private Integer numeroInteresados;
 
 }
